--- a/webpack.config.js
+++ b/webpack.config.js
@@ -38,6 +38,7 @@ module.exports = {
       ALGOLIA_PROJECT_ID: JSON.stringify(process.env.ALGOLIA_PROJECT_ID),
       ALGOLIA_PUBLIC_KEY: JSON.stringify(process.env.ALGOLIA_PUBLIC_KEY),
       ALGOLIA_INDEX: JSON.stringify(process.env.ALGOLIA_INDEX),
-    })
+      ENV: JSON.stringify(process.env.NODE_ENV),
+    }),
   ],
 };